backend/requirements.txt
Flask==3.0.3
gunicorn==22.0.0
google-auth==2.34.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.141.0
requests==2.32.3
